iommu/ipmmu-vmsa: Add Renesas R8A779F0 (R-Car S4) support
Based on the following commit from the Renesas BSP:
7003b9f732cffdc778fceb4bffb05ebb4540f726
located at:
https://github.com/renesas-rcar/linux-bsp/tree/v5.10.41/rcar-5.1.3.rc5

Original commit message:
 commit 7003b9f732cffdc778fceb4bffb05ebb4540f726
 Author: Nam Nguyen <nam.nguyen.yh@renesas.com>
 Date:   Thu Feb 4 11:05:37 2021 +0700

  iommu/ipmmu-vmsa: Add Renesas R8A779F0 (R-Car S4) support

  Adding IPMMU support for Renesas R8A779F0 R-Car S4.

  Suggested by: Hai Pham <hai.pham.ud@renesas.com>
  Signed-off-by: Nam Nguyen <nam.nguyen.yh@renesas.com>

**********

The R-Car S4 is an automotive System-on-Chip (SoC) for Car
Server/Communication Gateway and is one of the first products
in Renesas’ 4th-generation R-Car Family.

The integrated IOMMU HW is also VMSA-compatible and supports
stage 2 translation table format, therefore can be used with
current driver with slight modifications (thanks to the prereq
work).

In the context of Xen driver the main differences between Gen3
and S4 are the following:
- HW capacity was enlarged to support up to 16 IPMMU contexts
  (sets of page table) and up to 64 micro-TLBs per IPMMU device
- the memory mapped registers have different bases and offsets

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>
Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>