libxl/arm: Add basic virtio-iommu support
Introduce "virtio-iommu" config option (disabled by default) which
purpose is to indicate that virtio devices which serve a particular
guest are behind the IOMMU.
If set, generate virtio-iommu DT node and link all virtio devices
to it. We are going to use scenario where all virtio devices inside
a guest are connected to the same virtio-iommu unit (share the IOVA
space).

The virtio-iommu parameters (enable, MMIO irq and base) are passed
to the hypervisor via XEN_DOMCTL_createdomain using sub-arch fields.

The following setting must be enabled in the guest Linux defconfig:
CONFIG_VIRTIO_IOMMU=y

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>