[hack] xen/arm: Extend foreign memory mappings restriction for hardware domain
For the test purposes only. First of all there is no point for doing
that for the virtio backend running in trusted domain (we trust
the hardware domain by default). The second is that added check is
not precise. It is based on the wrong assumption that all foreign mapping
requests after domain creation will "always" be originated by the virtio
backend (during domain creation the toolstack is the origin of these
requests). But the foreign mappings can also be used by other entities
(some PV drivers, some version of Xenstore, etc).

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>