libxl: Check for d_config to be valid in device_model_spawn_outcome()
Check that d_config is not NULL before dereferencing it. For instance,
this is a case when we use Qemu in non toolstack domain for the sake
of Qdisk backend, so xl devd (which spawns Qemu) fails with segmentation
fault afterwards.

See libxl__spawn_qdisk_backend() which explicitly sets:
dmss->guest_config = NULL;

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>